refactoring functions to make it more reusable :D

+++ script.js
@@ -1,59 +1,69 @@
 let myLeads = [];
+let oldLeads = [1,2,3,4,5];
 const inputEl = document.querySelector("#input-el");
 const inputBtn = document.querySelector("#input-btn");
-const deleteBtn = document.querySelector('#delete-btn');
+const deleteBtn = document.querySelector("#delete-btn");
 const ulEl = document.querySelector("#ul-el");
 
-
-// Get the leads from the localStorage - PS: JSON.parse()
+// turn the localStorage into object/Array and Get the leads from the localStorage
 const leadsFromLocalStorage = JSON.parse(localStorage.getItem("myLeads"));
 console.log(leadsFromLocalStorage);
 
 //if may laman (true), if so then render the leads
- if(leadsFromLocalStorage) {
-  myLeads = leadsFromLocalStorage
-  renderLeads();
- }
+if (leadsFromLocalStorage) {
+  myLeads = leadsFromLocalStorage;
+  render(myLeads);
+}
 
- //deleting localStorage, Array, and DOM
-deleteBtn.addEventListener('click', function() {
-  localStorage.clear()
-  myLeads = [];
-  renderLeads();
-})
 
+
+//function to render the value of arrays with argument 'leads' to be more dynamic and reusable
+function render(leads) {
+  let listItems = "";
+  for (let i = 0; i < leads.length; i++) {
+    // 2. Add the item to the listItems variable
+    listItems += `<li><a  href="${leads[i]}" target='_blank' >${leads[i]}</a></li>`;
+  }
+  ulEl.innerHTML = listItems;
+}
+
+
+
+//deleting localStorage, Array, and DOM
+deleteBtn.addEventListener("dblclick", function () {
+  //dialog for confirmation before deleting all the leads
+  confirm("Are you sure you want to delete all the leads?");
+
+  if (confirm) {
+    //clear 
+    localStorage.clear();
+    myLeads = [];
+
+    //after deleting, call the function to render the leads
+    render(myLeads);
+  } else {
+    //if the user click cancel, then return to the page without deleting
+    return;
+  }
+});
 
 //refering to the input-btn
 inputBtn.addEventListener("click", function () {
   //pushing values from the inputs to the array 'myLeads'
 
-
-  //new variable that refering to the value of input
-  const currentInput = inputEl.value;
-
   //if not empty input, then push
-  if (currentInput.trim() !== "") {
-    myLeads.push(currentInput);
+  if (inputEl.value.trim() !== "") {
+    myLeads.push(inputEl.value.trim());
   }
 
   //clear inputs
   inputEl.value = "";
 
-  // Save the myLeads array to localStorage and transform the array to string -  
+  // Save the myLeads array to localStorage and transform the array to string -
   // within this statement to store the items in arrays as string
   localStorage.setItem("myLeads", JSON.stringify(myLeads));
 
   //calling the function to render the arrays when it's click
-  renderLeads();
+  render(myLeads);
 });
-
-//function to render the value of arrays
-function renderLeads() {
-  let listItems = "";
-  for (let i = 0; i < myLeads.length; i++) {
-    // 2. Add the item to the listItems variable
-    listItems += `<li><a  href="${myLeads[i]}" target='_blank' >${myLeads[i]}</a></li>`;
-  }
-  ulEl.innerHTML = listItems;
-}
